Normalise team search
